+++ app/components/game/Players.tsx
@@ -1,6 +1,7 @@
 import messages from "~/components/i18n/messages";
 import type { GameWithFeedback } from "~/routes/games/$gameId";
 import PlayerCard from "~/components/game/players/PlayerCard";
+import { Link } from "@remix-run/react";
 
 
 interface PlayerProps {
@@ -14,8 +15,8 @@ const Players = ({ game }: PlayerProps) => {
       <div className={"flex items-center gap-3"}>
         <p className={"text-gray-600 font-inter-regular text-subheading px-2 "}>{messages.game.players.title}</p>
         <div>
-          <button
-            className={"font-inter-regular text-item-caption bg-indigo-500 text-white rounded-lg py-2 px-3"}>{messages.game.players.add}</button>
+          <Link to={"player/create"}
+            className={"font-inter-regular text-item-caption bg-indigo-500 text-white rounded-lg py-2 px-3"}>{messages.game.players.add}</Link>
         </div>
       </div>
