+++ app/routes/games/$gameId/player/index.tsx
@@ -1,0 +1,8 @@
+
+const Index = () => {
+  return (
+    <div>Player Index</div>
+  );
+};
+
+export default Index;
