+++ prisma/schema.prisma
@@ -36,13 +36,13 @@ model Feedback {
  id    String @id @default(cuid())
  game Game @relation(fields: [gameId], references: [id])
  gameId String
- player Player @relation (fields: [playerId], references: [id])
+ player Index @relation (fields: [playerId], references: [id])
  playerId String
  status Boolean?
  note String?
 }
 
-model Player {
+model Index {
 id String @id @default(cuid())
 name String
 email String
